[FIX] Rename Component: CarouselCard
