+++ src/components/CarouselCard.js
